+++ core/interfaces.go
@@ -18,6 +18,9 @@ import (
 	"context"
 	"notifications/core/model"
 	"notifications/driven/storage"
+	"time"
+
+	"github.com/rokwire/logging-library-go/v2/logs"
 )
 
 // Services exposes APIs for the driver adapters
@@ -33,14 +36,20 @@ type Services interface {
 	UpdateUserByID(orgID string, appID string, userID string, notificationsEnabled bool) (*model.User, error)
 	DeleteUserWithID(orgID string, appID string, userID string) error
 
-	GetMessages(orgID string, appID string, userID *string, read *bool, mute *bool, messageIDs []string, startDateEpoch *int64, endDateEpoch *int64, filterTopic *string, offset *int64, limit *int64, order *string) ([]model.Message, error)
+	GetMessagesRecipientsDeep(orgID string, appID string, userID *string, read *bool, mute *bool, messageIDs []string, startDateEpoch *int64, endDateEpoch *int64, filterTopic *string, offset *int64, limit *int64, order *string) ([]model.MessageRecipient, error)
+
 	GetMessagesStats(orgID string, appID string, userID string) (*model.MessagesStats, error)
 	GetMessage(orgID string, appID string, ID string) (*model.Message, error)
-	CreateMessage(user *model.CoreUserRef, message *model.Message, async bool) (*model.Message, error)
+	GetUserMessage(orgID string, appID string, ID string, accountID string) (*model.Message, error)
+	CreateMessage(orgID string, appID string,
+		sender model.Sender, time time.Time, priority int, subject string, body string, data map[string]string,
+		inputRecipients []model.MessageRecipient, recipientsCriteriaList []model.RecipientCriteria,
+		recipientAccountCriteria map[string]interface{}, topic *string, async bool) (*model.Message, error)
 	UpdateMessage(userID *string, message *model.Message) (*model.Message, error)
 	DeleteUserMessage(orgID string, appID string, userID string, messageID string) error
 	DeleteMessage(orgID string, appID string, ID string) error
 	UpdateReadMessage(orgID string, appID string, ID string, userID string) (*model.Message, error)
+	UpdateAllUserMessagesRead(orgID string, appID string, userID string, read bool) error
 
 	GetAllAppVersions(orgID string, appID string) ([]model.AppVersion, error)
 	GetAllAppPlatforms(orgID string, appID string) ([]model.AppPlatform, error)
@@ -80,8 +89,8 @@ func (s *servicesImpl) UpdateTopic(topic *model.Topic) (*model.Topic, error) {
 	return s.app.updateTopic(topic)
 }
 
-func (s *servicesImpl) GetMessages(orgID string, appID string, userID *string, read *bool, mute *bool, messageIDs []string, startDateEpoch *int64, endDateEpoch *int64, filterTopic *string, offset *int64, limit *int64, order *string) ([]model.Message, error) {
-	return s.app.getMessages(orgID, appID, userID, read, mute, messageIDs, startDateEpoch, endDateEpoch, filterTopic, offset, limit, order)
+func (s *servicesImpl) GetMessagesRecipientsDeep(orgID string, appID string, userID *string, read *bool, mute *bool, messageIDs []string, startDateEpoch *int64, endDateEpoch *int64, filterTopic *string, offset *int64, limit *int64, order *string) ([]model.MessageRecipient, error) {
+	return s.app.getMessagesRecipientsDeep(orgID, appID, userID, read, mute, messageIDs, startDateEpoch, endDateEpoch, filterTopic, offset, limit, order)
 }
 
 func (s *servicesImpl) GetMessagesStats(orgID string, appID string, userID string) (*model.MessagesStats, error) {
@@ -92,8 +101,16 @@ func (s *servicesImpl) GetMessage(orgID string, appID string, ID string) (*model
 	return s.app.getMessage(orgID, appID, ID)
 }
 
-func (s *servicesImpl) CreateMessage(user *model.CoreUserRef, message *model.Message, async bool) (*model.Message, error) {
-	return s.app.createMessage(user, message, async)
+func (s *servicesImpl) GetUserMessage(orgID string, appID string, ID string, accountID string) (*model.Message, error) {
+	return s.app.getUserMessage(orgID, appID, ID, accountID)
+}
+
+func (s *servicesImpl) CreateMessage(orgID string, appID string,
+	sender model.Sender, time time.Time, priority int, subject string, body string, data map[string]string,
+	inputRecipients []model.MessageRecipient, recipientsCriteriaList []model.RecipientCriteria,
+	recipientAccountCriteria map[string]interface{}, topic *string, async bool) (*model.Message, error) {
+	return s.app.createMessage(orgID, appID, sender, time, priority, subject, body, data,
+		inputRecipients, recipientsCriteriaList, recipientAccountCriteria, topic, async)
 }
 
 func (s *servicesImpl) UpdateMessage(userID *string, message *model.Message) (*model.Message, error) {
@@ -102,6 +119,10 @@ func (s *servicesImpl) UpdateMessage(userID *string, message *model.Message) (*m
 
 func (s *servicesImpl) UpdateReadMessage(orgID string, appID string, ID string, userID string) (*model.Message, error) {
 	return s.app.updateReadMessage(orgID, appID, ID, userID)
+}
+
+func (s *servicesImpl) UpdateAllUserMessagesRead(orgID string, appID string, userID string, read bool) error {
+	return s.app.updateAllUserMessagesRead(orgID, appID, userID, read)
 }
 
 func (s *servicesImpl) DeleteUserMessage(orgID string, appID string, userID string, messageID string) error {
@@ -136,38 +157,85 @@ func (s *servicesImpl) SendMail(toEmail string, subject string, body string) err
 	return s.app.sendMail(toEmail, subject, body)
 }
 
+// BBs exposes users related APIs used by the platform building blocks
+type BBs interface {
+	BBsCreateMessage(orgID string, appID string,
+		sender model.Sender, time time.Time, priority int, subject string, body string, data map[string]string,
+		inputRecipients []model.MessageRecipient, recipientsCriteriaList []model.RecipientCriteria,
+		recipientAccountCriteria map[string]interface{}, topic *string, async bool) (*model.Message, error)
+	BBsDeleteMessage(l *logs.Log, serviceAccountID string, messageID string) error
+	BBsSendMail(toEmail string, subject string, body string) error
+}
+
+type bbsImpl struct {
+	app *Application
+}
+
+func (s *bbsImpl) BBsCreateMessage(orgID string, appID string,
+	sender model.Sender, time time.Time, priority int, subject string, body string, data map[string]string,
+	inputRecipients []model.MessageRecipient, recipientsCriteriaList []model.RecipientCriteria,
+	recipientAccountCriteria map[string]interface{}, topic *string, async bool) (*model.Message, error) {
+	return s.app.bbsCreateMessage(orgID, appID, sender, time, priority, subject, body, data,
+		inputRecipients, recipientsCriteriaList, recipientAccountCriteria, topic, async)
+}
+
+func (s *bbsImpl) BBsDeleteMessage(l *logs.Log, serviceAccountID string, messageID string) error {
+	return s.app.bbsDeleteMessage(l, serviceAccountID, messageID)
+}
+
+func (s *bbsImpl) BBsSendMail(toEmail string, subject string, body string) error {
+	return s.app.bbsSendMail(toEmail, subject, body)
+}
+
 // Storage is used by core to storage data - DB storage adapter, file storage adapter etc
 type Storage interface {
 	RegisterStorageListener(storageListener storage.Listener)
 
+	PerformTransaction(func(context storage.TransactionContext) error, int64) error
+
 	LoadFirebaseConfigurations() ([]model.FirebaseConf, error)
 
+	FindUsersByIDs(usersIDs []string) ([]model.User, error)
 	FindUserByID(orgID string, appID string, userID string) (*model.User, error)
 	UpdateUserByID(orgID string, appID string, userID string, notificationsEnabled bool) (*model.User, error)
 	DeleteUserWithID(orgID string, appID string, userID string) error
 
 	FindUserByToken(orgID string, appID string, token string) (*model.User, error)
 	StoreFirebaseToken(orgID string, appID string, tokenInfo *model.TokenInfo, userID string) error
-	GetFirebaseTokensByRecipients(orgID string, appID string, recipient []model.Recipient, criteriaList []model.RecipientCriteria) ([]string, error)
-	GetRecipientsByTopic(orgID string, appID string, topic string) ([]model.Recipient, error)
-	GetRecipientsByRecipientCriterias(orgID string, appID string, recipientCriterias []model.RecipientCriteria) ([]model.Recipient, error)
+	GetFirebaseTokensByRecipients(orgID string, appID string, recipient []model.MessageRecipient, criteriaList []model.RecipientCriteria) ([]string, error)
+	GetUsersByTopicWithContext(ctx context.Context, orgID string, appID string, topic string) ([]model.User, error)
+	GetUsersByRecipientCriteriasWithContext(ctx context.Context, orgID string, appID string, recipientCriterias []model.RecipientCriteria) ([]model.User, error)
 	SubscribeToTopic(orgID string, appID string, token string, userID string, topic string) error
 	UnsubscribeToTopic(orgID string, appID string, token string, userID string, topic string) error
 	GetTopics(orgID string, appID string) ([]model.Topic, error)
 	InsertTopic(*model.Topic) (*model.Topic, error)
 	UpdateTopic(*model.Topic) (*model.Topic, error)
 
-	GetMessages(orgID string, appID string, userID *string, read *bool, mute *bool, messageIDs []string, startDateEpoch *int64, endDateEpoch *int64, filterTopic *string, offset *int64, limit *int64, order *string) ([]model.Message, error)
+	FindMessagesRecipients(orgID string, appID string, messageID string, userID string) ([]model.MessageRecipient, error)
+	FindMessagesRecipientsDeep(orgID string, appID string, userID *string, read *bool, mute *bool, messageIDs []string, startDateEpoch *int64, endDateEpoch *int64, filterTopic *string, offset *int64, limit *int64, order *string) ([]model.MessageRecipient, error)
+	InsertMessagesRecipientsWithContext(ctx context.Context, items []model.MessageRecipient) error
+	DeleteMessagesRecipientsForMessageWithContext(ctx context.Context, messageID string) error
+
+	FindMessageWithContext(ctx context.Context, ID string) (*model.Message, error)
 	GetMessage(orgID string, appID string, ID string) (*model.Message, error)
-	CreateMessage(message *model.Message) (*model.Message, error)
+	CreateMessageWithContext(ctx context.Context, message model.Message) (*model.Message, error)
 	UpdateMessage(message *model.Message) (*model.Message, error)
 	DeleteUserMessageWithContext(ctx context.Context, orgID string, appID string, userID string, messageID string) error
 	DeleteMessageWithContext(ctx context.Context, orgID string, appID string, ID string) error
-	GetMessagesStats(userID string, read bool, mute bool) (*model.MessagesStats, error)
+	GetMessagesStats(userID string) (*model.MessagesStats, error)
 	UpdateUnreadMessage(ctx context.Context, orgID string, appID string, ID string, userID string) (*model.Message, error)
-
+	UpdateAllUserMessagesRead(ctx context.Context, orgID string, appID string, userID string, read bool) error
 	GetAllAppVersions(orgID string, appID string) ([]model.AppVersion, error)
 	GetAllAppPlatforms(orgID string, appID string) ([]model.AppPlatform, error)
+
+	InsertQueueDataItemsWithContext(ctx context.Context, items []model.QueueItem) error
+	LoadQueueWithContext(ctx context.Context) (*model.Queue, error)
+	SaveQueueWithContext(ctx context.Context, queue model.Queue) error
+	SaveQueue(queue model.Queue) error
+
+	FindQueueData(time *time.Time, limit int) ([]model.QueueItem, error)
+	DeleteQueueData(ids []string) error
+	DeleteQueueDataForMessageWithContext(ctx context.Context, messageID string) error
 }
 
 // Firebase is used to wrap all Firebase Messaging API functions
@@ -182,4 +250,9 @@ type Firebase interface {
 // Mailer is used to wrap all Email Messaging functions
 type Mailer interface {
 	SendMail(toEmail string, subject string, body string) error
+}
+
+// Core exposes Core APIs for the driver adapters
+type Core interface {
+	RetrieveCoreUserAccountByCriteria(accountCriteria map[string]interface{}, appID *string, orgID *string) ([]model.CoreAccount, error)
 }